+++ concordium-consensus/src/Concordium/GlobalState/Persistent/BlockState.hs
@@ -2768,14 +2768,15 @@ doIsProtocolUpdateEffective :: (SupportsPersistentState pv m) => PersistentBlock
 doIsProtocolUpdateEffective = isProtocolUpdateEffective . bspUpdates <=< loadPBS
 
 doProcessUpdateQueues ::
+    forall pv m.
     (SupportsPersistentState pv m) =>
     PersistentBlockState pv ->
     Timestamp ->
-    m (Map.Map TransactionTime (UpdateValue (ChainParametersVersionFor pv)), PersistentBlockState pv)
+    m ([(TransactionTime, UpdateValue (ChainParametersVersionFor pv))], PersistentBlockState pv)
 doProcessUpdateQueues pbs ts = do
     bsp <- loadPBS pbs
     let (u, ars, ips) = (bspUpdates bsp, bspAnonymityRevokers bsp, bspIdentityProviders bsp)
-    (changes, (u', ars', ips')) <- processUpdateQueues ts (u, ars, ips)
+    (changes, (u', ars', ips')) <- processUpdateQueues (protocolVersion @pv) ts (u, ars, ips)
     (changes,) <$> storePBS pbs bsp{bspUpdates = u', bspAnonymityRevokers = ars', bspIdentityProviders = ips'}
 
 doProcessReleaseSchedule :: forall m pv. (SupportsPersistentState pv m) => PersistentBlockState pv -> Timestamp -> m (PersistentBlockState pv)